Fix: World 변경 로직 수정

+++ backend/src/main/java/com/vvs/peekpick/member/controller/MemberController.java
@@ -167,7 +167,7 @@ public class MemberController {
 
     @PostMapping("/world")
     public CommonResponse updateWorld(Authentication authentication,
-                                      @RequestBody Map<String, Integer> param) {
+                                      @RequestBody Map<String, Long> param) {
         Long avatarId = Long.valueOf(authentication.getName());
         memberService.updateWorld(avatarId, param.get("worldId"));
